refactor(platform): expose Users on its own /users subpath
Consumers who only need user management should not pull in the whole
platform module. Users moves to `@uipath/uipath-typescript/users`,
mirroring the per-subpath layout used by orchestrator/assets and
maestro/cases: the barrel lives beside the service, models stay flat
under src/models/platform.

`/platform` now exports only the settings service and its models.
