+++ src/services/platform/index.ts
@@ -1,17 +1,17 @@
 /**
  * Platform Module
  *
- * Provides access to UiPath platform administration:
+ * Provides access to UiPath platform settings:
  * - `Platform` — bulk read and bulk create/update of a user's setting key/value pairs
- * - `Users` — list, read, and update an organization's users, including group membership
  *
- * Settings operations are user-scoped — see {@link PlatformSettingKey}. Settings require
- * the `PM.Setting` scope (or `PM.Setting.Read` / `PM.Setting.Write`).
+ * Every operation is user-scoped — see {@link PlatformSettingKey}.
+ *
+ * Requires the `PM.Setting` scope (or `PM.Setting.Read` / `PM.Setting.Write`).
  *
  * @example
  * ```typescript
  * import { UiPath } from '@uipath/uipath-typescript/core';
- * import { Platform, PlatformSettingKey, Users } from '@uipath/uipath-typescript/platform';
+ * import { Platform, PlatformSettingKey } from '@uipath/uipath-typescript/platform';
  *
  * const sdk = new UiPath(config);
  * await sdk.initialize();
@@ -26,16 +26,13 @@
  *   settings[0].userId,
  *   settings[0].organizationId
  * );
- *
- * const users = new Users(sdk);
- * const allUsers = await users.getAll('<organizationId>');
  * ```
  *
  * @module
  */
 
 export { PlatformService as Platform } from './platform';
-export { PlatformUserService as Users } from './users';
 
 // Models (types, response shapes)
-export * from '../../models/platform';
+export * from '../../models/platform/platform.types';
+export * from '../../models/platform/platform.models';
